front: internal support for file for CSV upsert
